refactor(pages): :bricks: add common layout for podcast and episode pages
