+++ src/pages/PodcastEpisodePage.tsx
@@ -1,29 +1,19 @@
 import loadable from '@loadable/component'
 import { atom, useAtom } from 'jotai'
 import { memo, useEffect, useMemo } from 'react'
-import { useParams } from 'react-router-dom'
+import { useOutletContext } from 'react-router-dom'
 
-import ContainerTransition from '~/components/routing/atoms/ContainerTransition'
-import DetailsCard from '~/components/ui/molecules/DetailsCard'
-import usePodcast from '~/hooks/query/usePodcast'
-import usePodcastInfo from '~/hooks/query/usePodcastInfo'
-import type { Item } from '~/interfaces/PodcastInfo'
+import type { Channel, Item } from '~/interfaces/PodcastInfo'
 
 export const episodeSelectedAtom = atom<Item | null>(null)
 export const episodeNotFoundAtom = atom<boolean>(false)
 
 function PodcastEpisodePage(): React.ReactElement {
-  const { podcastId = '', episodeId = '' } = useParams<{
-    podcastId: string | undefined
-    episodeId: string | undefined
+  const { podcastInfo, episodeId, podcastId } = useOutletContext<{
+    podcastInfo: Channel | undefined
+    episodeId: string
+    podcastId: string
   }>()
-
-  const { data: podcast } = usePodcast(podcastId)
-
-  const { data: podcastInfo } = usePodcastInfo(
-    podcastId,
-    podcast?.feedUrl ?? ''
-  )
 
   const [episode, setEpisode] = useAtom(episodeSelectedAtom)
   const [episodeNotFound, setEpisodeNotFoundAtom] = useAtom(episodeNotFoundAtom)
@@ -39,24 +29,6 @@ function PodcastEpisodePage(): React.ReactElement {
       setEpisodeNotFoundAtom(true)
     }
   }, [podcastInfo, episodeId])
-
-  const DetailsCardMemo = useMemo(
-    () => (
-      <DetailsCard
-        {...(podcast != null && {
-          image: {
-            src: podcast.artworkUrl600,
-            alt: podcast.collectionName
-          },
-          title: podcast.collectionName,
-          author: podcast.artistName
-        })}
-        link={{ to: `/podcast/${podcastId}` }}
-        descriptionInnerHTML={podcastInfo?.description}
-      />
-    ),
-    [podcast, podcastInfo]
-  )
 
   const ContentLoadingMemo = useMemo(() => {
     return <span className='loading btn-ghost btn-lg btn transition-all'></span>
@@ -103,12 +75,9 @@ function PodcastEpisodePage(): React.ReactElement {
   )
 
   return (
-    <ContainerTransition className='container mx-auto flex w-full flex-col items-center gap-3 px-4 pt-8 pb-9 md:flex-row md:items-start lg:gap-24'>
-      {DetailsCardMemo}
-      <div className='flex w-full flex-col gap-5 border border-base-200 bg-base-100 p-5 drop-shadow-lg'>
-        {ContentFoundRenderMemo}
-      </div>
-    </ContainerTransition>
+    <div className='flex w-full flex-col gap-5 border border-base-200 bg-base-100 p-5 drop-shadow-lg'>
+      {ContentFoundRenderMemo}
+    </div>
   )
 }
