+++ src/pages/PodcastPage.tsx
@@ -1,28 +1,19 @@
 import dayjs from 'dayjs'
 import { memo, useCallback, useMemo } from 'react'
-import { useNavigate, useParams } from 'react-router-dom'
+import { useNavigate, useOutletContext } from 'react-router-dom'
 
-import ContainerTransition from '~/components/routing/atoms/ContainerTransition'
 import PodcastEpisodesCount from '~/components/ui/atoms/PodcastEpisodesCount'
-import DetailsCard from '~/components/ui/molecules/DetailsCard'
 import InfoTable from '~/components/ui/molecules/InfoTable'
-import usePodcast from '~/hooks/query/usePodcast'
-import usePodcastInfo from '~/hooks/query/usePodcastInfo'
+import type { Channel } from '~/interfaces/PodcastInfo'
 
 const intl = new Intl.DateTimeFormat('en-US')
 
 function PodcastPage(): React.ReactElement {
-  const nav = useNavigate()
-  const { podcastId = '' } = useParams<{
-    podcastId: string | undefined
+  const { podcastInfo } = useOutletContext<{
+    podcastInfo: Channel | undefined
   }>()
 
-  const { data: podcast } = usePodcast(podcastId)
-
-  const { data: podcastInfo } = usePodcastInfo(
-    podcastId,
-    podcast?.feedUrl ?? ''
-  )
+  const nav = useNavigate()
 
   const handleSelected = useCallback(
     (guid: string) => {
@@ -34,20 +25,6 @@ function PodcastPage(): React.ReactElement {
   const podcastInfoLength = useMemo(
     () => podcastInfo?.item.length ?? 0,
     [podcastInfo]
-  )
-
-  const DetailsCardMemo = useMemo(
-    () => (
-      <DetailsCard
-        {...(podcast != null && {
-          title: podcast.collectionName,
-          author: podcast.artistName,
-          image: { src: podcast.artworkUrl600, alt: podcast.collectionName }
-        })}
-        descriptionInnerHTML={podcastInfo?.description}
-      />
-    ),
-    [podcast, podcastInfo]
   )
 
   const InfoTableMemo = useMemo(
@@ -72,15 +49,12 @@ function PodcastPage(): React.ReactElement {
   )
 
   return (
-    <ContainerTransition className='container mx-auto flex w-full flex-col items-center gap-3 px-4 pt-8 pb-9 md:flex-row md:items-start lg:gap-24'>
-      {DetailsCardMemo}
-      <div className='flex w-full flex-col gap-5'>
-        {PodcastEpisodesCountMemo}
-        <div className='w-full border border-base-200 bg-base-100 p-5 drop-shadow-lg'>
-          {InfoTableMemo}
-        </div>
+    <div className='flex w-full flex-col gap-5'>
+      {PodcastEpisodesCountMemo}
+      <div className='w-full border border-base-200 bg-base-100 p-5 drop-shadow-lg'>
+        {InfoTableMemo}
       </div>
-    </ContainerTransition>
+    </div>
   )
 }
